[dependency refactor] Refactor ibm wrapper to not require optional dependencies.
Summary:

Test Plan:
